events_page/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==21.4.0
    # via jsonlines
beautifulsoup4==4.10.0
    # via cloudflare
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via flask
cloudflare==2.8.15
    # via -r requirements.in
flask==2.0.2
    # via
    #   -r requirements.in
    #   flask-assets
    #   frozen-flask
flask-assets==2.0
    # via -r requirements.in
frozen-flask==0.18
    # via -r requirements.in
google-api-core[grpc]==2.3.2
    # via
    #   google-api-python-client
    #   google-cloud-build
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-api-python-client==2.33.0
    # via -r requirements.in
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==0.4.6
    # via -r requirements.in
google-cloud-build==3.7.1
    # via -r requirements.in
google-cloud-core==2.2.1
    # via google-cloud-storage
google-cloud-secret-manager==2.8.0
    # via -r requirements.in
google-cloud-storage==1.44.0
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.1.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.54.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.3
    # via google-cloud-secret-manager
grpcio==1.43.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.43.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
jsonlines==3.0.0
    # via cloudflare
libcst==0.3.23
    # via google-cloud-secret-manager
libsass==0.21.0
    # via -r requirements.in
logzero==1.7.0
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
mypy-extensions==0.4.3
    # via typing-inspect
oauthlib==3.1.1
    # via requests-oauthlib
proto-plus==1.19.8
    # via
    #   google-cloud-build
    #   google-cloud-secret-manager
protobuf==3.19.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.6
    # via httplib2
python-dateutil==2.8.2
    # via -r requirements.in
pyyaml==6.0
    # via
    #   cloudflare
    #   libcst
requests==2.26.0
    # via
    #   -r requirements.in
    #   cloudflare
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
ruamel.yaml==0.17.20
    # via -r requirements.in
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-storage
    #   grpcio
    #   libsass
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
typing-extensions==4.0.1
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.7
    # via requests
webassets==2.0
    # via flask-assets
werkzeug==2.0.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
